pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi[standard]>=0.136.1",
    "itsdangerous>=2.2.0",
    "pyjwt[crypto]>=2.10.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "loguru>=0.7.3",
    "rich>=15.0.0",
    "uvicorn[standard]>=0.46.0",
    "uvicorn-worker>=0.4.0",
    "gunicorn>=25.3.0",
    "sqladmin>=0.25.0",
    "alembic>=1.18.4",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg >=0.31.0",
    "psycopg[binary,pool]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "aiobotocore>=3.0.0",
    "opendal>=0.47.1",
    "python-magic>=0.4.27",
    "charset-normalizer>=3.4.7",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
addopts = "-s"
asyncio_mode = "auto"
pythonpath = "."
testpaths = ["tests", "app/apps/accounts/tests", "app/apps/examples/tests", "app/apps/file_manager/tests"]

[dependency-groups]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "aiosqlite>=0.22.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF", "ASYNC", "PT"]
ignore = [
    "F405",
    "RUF001", # ambiguous unicode in strings (Chinese punctuation is fine)
    "RUF002", # ambiguous unicode in docstrings
    "RUF003", # ambiguous unicode in comments
    "PT022",  # prefer uniform yield-style fixtures even without teardown
]

[tool.ruff.lint.per-file-ignores]
"app/apps/examples/api/param_examples.py" = ["ALL"]
"tests/integration/conftest.py" = ["SIM117"]

[tool.black]
line-length = 120
skip-string-normalization = true
